person_detection/requirements.txt
# Core
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
certifi>=2024.0.0

# YOLO models
ultralytics>=8.0.0

# Torchvision models (Faster R-CNN, SSD MobileNet)
torch>=2.0.0
torchvision>=0.15.0
